pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "legate-studio"
version = "0.3.0"
description = "Legato Memory Hub - Dashboard, Dropbox, and RAG API"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "bobbyhiddn"}
]
dependencies = [
    "flask>=3.0.0",
    "flask-limiter>=3.5.0",
    "requests>=2.31.0",
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.2.0",
    "boto3>=1.34.0",
    "markdown>=3.5.0",
    "pygments>=2.17.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "stripe>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
legate-studio = "legate_studio.main:main"
legate-recovery = "legate_studio.recovery:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
